Remove redundant class
